Moved createPage call to RCPageList from RCStoryEditor
