+++ lib/ReactViews/RCBuilder/RCPageList/RCPageList.jsx
@@ -4,6 +4,7 @@ import { listPages } from "../../../../api/graphql/queries";
 import * as mutations from "../../../../api/graphql/mutations";
 import { useParams, withRouter, useHistory } from "react-router-dom";
 import Styles from "./RCPageList.scss";
+import RCAccordian from "../RCAccordian/RCAccordian";
 import Icon from "../../Icon";
 import orderList from "./RCOrderList";
 
@@ -12,6 +13,9 @@ function RCPageList() {
   // get the story id from url
   const { id } = useParams();
   const history = useHistory();
+
+  // get all the pages for this story
+  // TODO: feth pages by story id
   useEffect(() => {
     try {
       API.graphql(graphqlOperation(listPages)).then(data => {
@@ -23,6 +27,34 @@ function RCPageList() {
       console.log(error);
     }
   }, []);
+  // Create Page
+  const addPage = () => {
+    const newPage = {
+      title: "New page",
+      section: "INTRODUCTION",
+      camera: [0, 0, 0, 0],
+      baseMapName: "basemap",
+      viewer_mode_3d: true,
+      scenarios: []
+    };
+
+    // Create a new page
+    API.graphql({
+      query: mutations.createPage,
+      variables: { input: newPage }
+    }).then(response => {
+      if (response.data.createPage) {
+        // Add the new page to the story
+        newPage.id = response.data.createPage.id;
+        const newPages = Array.isArray(pages) ? [...pages, newPage] : [newPage];
+        setPages(newPages);
+        console.log("pages saved", newPages);
+      } else {
+        console.log("Error", response.errors[0].message);
+      }
+    });
+  };
+  // Delete Page
   const deletePage = id => {
     try {
       // const deletedPage = API.graphql(graphqlOperation(deletePage, { input: { id } }));
@@ -39,8 +71,15 @@ function RCPageList() {
   const toEditPage = pageId => {
     history.push(`/builder/story/${id}/page/${pageId}/edit`);
   };
-  return pages
-    ? pages.map(page => {
+  return pages ? (
+    <RCAccordian
+      title="Pages"
+      hasAction={true}
+      actionTitle="+ Add"
+      action={addPage}
+      enableReorder={true}
+    >
+      {pages.map(page => {
         return (
           <li key={page.id} className={Styles.listItem}>
             <Icon glyph={Icon.GLYPHS.reorder} class="reorder" />
@@ -53,7 +92,8 @@ function RCPageList() {
             </button>
           </li>
         );
-      })
-    : null;
+      })}
+    </RCAccordian>
+  ) : null;
 }
 export default withRouter(RCPageList);
