+++ lib/ReactViews/RCBuilder/RCPageList/RCPageList.scss
@@ -4,7 +4,7 @@
   min-height: 30px;
   padding: 5px;
   margin-bottom: 12px;
-  background: #fafafa;
+  background: #f6f6fa;
   border-radius: 8px;
   display: grid;
   grid-template-columns: 40px 1fr 40px 40px;
